% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{align}
\usage{
align(x, method = "centroid")
}
\arguments{
\item{x}{An sf object, already processed with `converge`, to be transformed in alignment; REQUIRED.}

\item{method}{Method used to align "top", "bottom", "left", "right", "vcentre", "hcentre";
"vcentroid", "hcentroid", "centroid"; default="centroid".}
}
\value{
An sf object containing one or more features (with no defined CRS)
}
\description{
NOT YET IMPLEMENTED
This function aligns simple features spatially by bounding box or centroid values.
Bounding box alignment or justification types include: top, bottom, left, right, vcentre, hcentre.
Centroid alignment include: vcentroid, hcentroid, centroid
}
\details{
To visually distribute the resulting features, use the `distribute` function.
}
\seealso{
\code{\link{converge}}, \code{\link{normalize}}, \code{\link{scale}}, \code{\link{distribute}}
}
\keyword{cartogram,}
\keyword{infographic}
\keyword{sf,}
