% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{get_table_layout}
\alias{get_table_layout}
\title{get_table_layout}
\usage{
get_table_layout(
  n = 1,
  i = 1:n,
  origincorner = "topleft",
  direction = "row",
  cols = NULL,
  rows = NULL
)
}
\arguments{
\item{n}{The number of items, if vector is to start from 1 and end at n. Default=1}

\item{i}{A vector of numbers to be ordered. Default=1:n}

\item{direction}{The direction by which the table should be filled: row, column. Default="row"}

\item{cols}{Number of columns to define the matrix}

\item{rows}{Number of rows to define the matrix}

\item{originalcorner}{The corner from which the matrix should be filled: topleft, bottomleft, topright, bottomright. Default="bottomleft"}
}
\value{
A matrix containing the vector.

If only cols is present, rows are calculated automatically to make a matrix.
If neither cols or rows are present, cols is set to sqrt of length of vector.
The vector is also padded with NA to create empty cells where
the cols*rows is greater than the length of the initial vector.
}
\description{
This helper function converts a single-dimensional vector into a two-dimensional
tabular matrix.
}
\details{
In morphogram, it is used to help structure the tabular distribution layouts.
Optional are the origin corner, direction of filling, number columns and/or rows.
}
