% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{get_table_layout_index}
\alias{get_table_layout_index}
\title{get_table_layout_index}
\usage{
get_table_layout_index(layout, i = 1:(nrow(layout) * ncol(layout)))
}
\description{
An internal convenience function to transform layout matrix into a data.frame, listing row, col
coordinates according to the unique numerical vector i, default being 1:ncells in layout
}
