% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{get_transform}
\alias{get_transform}
\title{get_transform}
\usage{
get_transform(x, method = "regulargrid", ...)
}
\arguments{
\item{x}{An sf-compatible feature layer, often containing polygons whose size is to be visually compared; REQUIRED.}

\item{method}{Method used to distribute features; default and only functioning method is "regulargrid".

Other parameters used depend on the transformation method used.}

\item{direction}{Direction of overall diagram, "row" fills each row first, "column" fills each column first; default "row".}

\item{cols}{Number of columns of features to plot before moving onto next line if dir="v".}

\item{rows}{Number of rows of features to plot before moving onto next line if dir="h".}

\item{margin}{Scalar coefficient of spacing between features; default 1.1.}

\item{x.mar}{Scalar coefficient of cell spacing on x axis, note that this is compounded on margin; default 1.}

\item{y.mar}{Scalar coefficient of cell spacing on y axis, note that this is compounded on margin; default 1.}

\item{x.nudge}{Additional spacing to add to x margins; default 0.}

\item{y.nudge}{Additional spacing to add to y margins; default 0.}
}
\description{
An internal function which calculates a set of transformations needed for a set of features and
returns these transformation instructions as a data.frame with columns xshift, yshift for affine transforms,
rotate for rotation transformations as an angle, plus xcellwidth, ycellheight for grid based methods
to help find the cell heights and widths per feature.
}
