% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphogram.R
\docType{package}
\name{morphogram}
\alias{morphogram}
\title{morphogram: A package to construct spatially-sensitive morphogram tableaux of spatial features.}
\description{
The morphogram package provides two main functions:
`converge`, and `distribute`, with additional support from `align`, `normalize` and `scale`.
}
\section{morphogram functions}{


Constructing a morphogram requires two or more steps:
1. transforming the spatial features so that they are centred around 0,0 of a
notional or mock projection using metres - `converge`;
2. (optionally) transforming the converged features to particular alignment,
or scale - `align`, `normalize` and `scale`
3. re-distributing features across the mock projection so that they can be
plotted discretely and compared visually - `distribute`.

Please read the Getting Started vignette ('morphogram') for more information.
}

