% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x, target = 1, method = "area")
}
\arguments{
\item{x}{An sf object, already processed with `converge`, to be transformed to normalized scale; REQUIRED.}

\item{target}{Target value to be normalized to, usually a single numeric, but can also be a vector of numeric target values of equal length to number of features in x; value meaning depends on method, e.g. square metres for areal, or metres for linear methods; default=1.}

\item{method}{Method of scaling: "area", "width", "height", "maxlinear"; default="area".}
}
\value{
An sf object containing one or more features (with no defined CRS)
}
\description{
NOT YET IMPLEMENTED
This function normalizes the scale of sf features, normally polygons. By default the scaled parameter
is the polygon *area*, so that each feature has an area of `scale` m2, whatever the actual shape of the polygon.
}
\details{
Features should have first been converged into a single comparable scale, using the `converge` function.
}
\seealso{
\code{\link{converge}}, \code{\link{align}}, \code{\link{normalize}}, \code{\link{scale}}, \code{\link{distribute}}
}
\keyword{cartogram,}
\keyword{infographic}
\keyword{sf,}
