% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeAroundCentroid.R
\name{normalizeAroundCentroid}
\alias{normalizeAroundCentroid}
\title{normalizeAroundCentroid}
\usage{
normalizeAroundCentroid(x, by.feature = T, combine = F)
}
\arguments{
\item{x}{An sf-compatible feature layer, often containing polygons whose size is to be visually compared; REQUIRED.}

\item{by.feature}{Whether to reproject by single features by individual feature centroids, T,
or reproject all features by a single centroid of the union of all features, F; default=TRUE.}

\item{combine}{Combine multiple geometries into one; default=F.}
}
\value{
An sf object containing one or more features (with no defined CRS)
}
\description{
This function finds the centroid of a set of sf-based features and reprojects them so that each feature's
centroid lies at the centre (0,0) of an arbitrary metre-based mercator projection. The resulting featuresh
are thus projected overlapping at the same scale, which can be visually compared. Note that the resulting
sf object has no CRS.
}
\details{
To visually distribute the resulting features, use the `distribute` function.
}
\seealso{
\code{\link{distribute}}
}
\keyword{cartogram,}
\keyword{infographic}
\keyword{sf,}
