% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotate}
\usage{
rotate(x, angle = c(0), origin = "centroid")
}
\arguments{
\item{x}{An sf object, to be transformed in alignment; REQUIRED.}

\item{angle}{Angle in degrees; default=0.}

\item{origin}{Whether to use centroid or coordinates origin; default="centroid".}
}
\value{
An sf object containing one or more features
}
\description{
This function rotates simple features spatially by centroid or origin.
}
\seealso{
\code{\link{converge}}, \code{\link{align}}, \code{\link{normalize}}, \code{\link{scale}}, \code{\link{distribute}}
}
\keyword{rotation}
\keyword{sf,}
