% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(x, scale = 1, method = "area")
}
\arguments{
\item{x}{An sf object, already processed with `converge`, to be transformed in scale; REQUIRED.}

\item{scale}{Scaling factor to be used, either a single numeric or a vector of numeric scales of equal length to number of features in x; default=1.}

\item{method}{Method of scaling: "area", "width", "height", "maxlinear"; default="area".}
}
\value{
An sf object containing one or more features (with no defined CRS)
}
\description{
NOT YET IMPLEMENTED
This function scales the size of sf features, by either linear or areal amounts.
}
\details{
Features should have first been converged into a single comparable scale, using
the `converge` function, and usually normalized in area or linear dimensions using
`normalize`.
}
\seealso{
\code{\link{converge}}, \code{\link{align}}, \code{\link{normalize}},  \code{\link{distribute}}
}
\keyword{cartogram,}
\keyword{infographic}
\keyword{sf,}
