% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcompare.R
\docType{package}
\name{spatialcompare}
\alias{spatialcompare}
\alias{spatialcompare-package}
\title{spatialcompare: A package to construct spatially-sensitive tableaux of spatial features.}
\description{
The spatialcompare package provides two main functions:
normalizeAroundCentroid, and distribute.
}
\section{spatialcompare functions}{

Constructing a spatialcompare graph requires two steps: 1. transforming the spatial features so
that they are centred around 0,0 of a notional or mock projection using metres;
2. re-distributing features across the mock projection so that they can be plotted discretely
and compared visually.
}

