% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{st_bbox_and_dims}
\alias{st_bbox_and_dims}
\title{st_bbox_and_dims}
\usage{
st_bbox_and_dims(x)
}
\arguments{
\item{x}{An sf-compatible feature layer REQUIRED.}
}
\value{
A data.frame with columns xmin, ymin, xmax, ymax, width, height
}
\description{
This helper function returns the bounding boxes and width/height dimensions of each feature row in an sf object.
}
