% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{distribute}
\usage{
distribute(
  x,
  preserve.parameters = T,
  max.features = 240,
  method = "regulargrid",
  label.points = F,
  label.pos = 1,
  ...
)
}
\arguments{
\item{x}{An sf-compatible feature layer, often containing polygons whose size is to be visually compared; REQUIRED.}

\item{preserve.parameters}{Whether to preserve non-geometry parameters; default TRUE (deprecated).}

\item{max.features}{Maximum features to compare; default=240.}

\item{method}{Method used to distribute features; default and only functioning method is "regulargrid".}

\item{label.points}{If TRUE, will add two columns label_x and label_y to resulting sf data.frame based on label.pos; default FALSE.}

\item{label.pos}{Position value for point at edge of regular grid squares, 0=centre, 1=top, 2=right, 3=bottom, 4=left; default=0.}
}
\value{
An sf object containing one or more features (with no defined CRS)

##
}
\description{
This function spatially distributes a set of sf features.
}
\details{
The spatial features should be aligned around 0,0 for meaningful results.
First use `converge` to bring together diversely distributed polygons around a single origin.
Use `align` to align different sets of simple spatial features around the origin.

Note that to rotate or scale features individually, you can use the `rotate` and `scale`
methods with a vector parameters.
}
\seealso{
\code{\link{converge}}
}
