% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{distribute}
\usage{
distribute(x, preserve.parameters = T, method = "regulargrid",
  cols = NULL, rows = NULL, dir = "v", max.features = 200,
  margin = 1.2, x.mar = 1, y.mar = 1, scale = 1, angle = 0)
}
\arguments{
\item{x}{An sf-compatible feature layer, often containing polygons whose size is to be visually compared; REQUIRED.}

\item{preserve.parameters}{Whether to preserve non-geometry parameters; default T.}

\item{method}{Method used to distribute features; default and only functioning method is "regulargrid".}

\item{cols}{Number of columns of features to plot before moving onto next line if dir="v".}

\item{rows}{Number of rows of features to plot before moving onto next line if dir="h".}

\item{dir}{Direction of overall diagram, "v" impiies filling each row first, "h" implies filling each column first; default "v".}

\item{max.features}{Maximum features to compare; default=Inf.}

\item{margin}{Scalar coefficient of spacing between features; default 1.2.}

\item{x.mar}{Scalar coefficient of spacing between x values; default 1.}

\item{y.mar}{Scalar coefficient of spacing between x values; default 1.}

\item{scale}{Affine scaling of feature dimensions; default 1.}

\item{angle}{Affine rotation of feature in degrees; default 0.}
}
\value{
An sf object containing one or more features (with no defined CRS)
}
\description{
This function spatially distributes a set of sf features.
}
\details{
The spatial features should be centred around 0,0. Use `align` to align different
sets of simple spatial features around 0,0.
}
\seealso{
\code{\link{converge}}
}
